I should have worried more about the overflow

+++ template/template.hpp
@@ -29,10 +29,10 @@ template <class T> using pque = priority_queue<T, vector<T>, greater<T>>;
 
 #endif
 
-#define rep(i, a, b) for (int i = (a); i < int(b); i++)
-#define REP(i, a, b) for (int i = (a); i <= int(b); i++)
-#define rrep(i, a, b) for (int i = int(b) - 1; i >= (a); i--)
-#define RREP(i, a, b) for (int i = int(b); i >= (a); i--)
+#define rep(i, a, b) for (ll i = (a); i < ll(b); i++)
+#define REP(i, a, b) for (ll i = (a); i <= ll(b); i++)
+#define rrep(i, a, b) for (ll i = ll(b) - 1; i >= (a); i--)
+#define RREP(i, a, b) for (ll i = ll(b); i >= (a); i--)
 #define fore(x, a) for (auto& x : (a))
 #define all(a) (a).begin(), (a).end()
 #define rall(a) (a).rbegin(), (a).rend()
